Moved GlueTabBar to glue.utils.qt.helpers
