+++ glue/utils/qt/colors.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function
 import numpy as np
 from matplotlib.colors import ColorConverter
 
-from ...external.qt import QtGui
+from glue.external.qt import QtGui
 
 from matplotlib import cm
