+++ glue/utils/qt/dialogs.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from ...external.qt import QtGui
+from glue.external.qt import QtGui
 
 # TODO: update docstrings
